+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improved error message output for objects that throw exceptions in their `toString()` methods. ([Issue 716](https://github.com/jqno/equalsverifier/issues/716))
 
+### Fixed
+
+- Fix `ClassFormatError` when attempting to create a dynamic subclass for a class that has no package. ([Issue 638](https://github.com/jqno/equalsverifier/issues/638))
+
 ## [3.12] - 2022-11-30
 
 ### Added